# mobile-sdk/openapi.yaml
openapi: 3.0.3
info:
  title: FrontFace Mobile SDK API
  version: "1.0.0"
  description: >
    Public API surface used by the FrontFace mobile chat SDK (Flutter). Covers AI chat,
    live human handoff, lead capture, and customer identify. See INTEGRATION_GUIDE.md for the
    behavioral contract (session lifecycle, realtime handoff, polling fallback, state machine).


    All endpoints are authenticated with a publishable client key in the `X-FrontFace-Key`
    header and identified by a stable per-install `X-Visitor-Id`. Live agent replies are
    delivered out-of-band over Supabase Realtime (channel `conversation:<id>`); that WebSocket
    transport is not described here — see the guide.
servers:
  - url: https://api.frontface.app
    description: Production
security:
  - FrontFaceKey: []
...
paths:
...
  /api/chat/message:
    post:
      tags: [Chat]
      summary: Send a message and get the AI reply (synchronous)
      operationId: sendMessage
      parameters:
        - $ref: "#/components/parameters/VisitorIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChatMessageRequest" }
      responses:
        "200":
          description: AI response (may include handoff signals)
          headers:
            X-RateLimit-Remaining: { schema: { type: integer }, description: Remaining requests in window }
            X-RateLimit-Reset: { schema: { type: integer }, description: Window reset (unix seconds) }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatMessageResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "504": { $ref: "#/components/responses/Timeout" }
  /api/chat/ensure-conversation:
    post:
      tags: [Chat]
      summary: Create a conversation up front (e.g. before triggering handoff with no message yet)
      operationId: ensureConversation
      parameters:
        - $ref: "#/components/parameters/VisitorIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, visitorId]
              properties:
                projectId: { type: string, format: uuid }
                visitorId: { type: string, maxLength: 100 }
                source: { $ref: "#/components/schemas/Source" }
      responses:
        "200":
          description: Conversation id
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversationId: { type: string, format: uuid }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    FrontFaceKey:
      type: apiKey
      in: header
      name: X-FrontFace-Key
      description: Publishable client key (pk_...). Safe to ship in the app binary; revocable.
  parameters:
    VisitorIdHeader:
      name: X-Visitor-Id
      in: header
      required: true
      description: Stable per-install visitor id (also mirrored into request bodies).
      schema: { type: string, maxLength: 100 }
...
  responses:
...
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Missing/invalid/revoked key, or key/project mismatch
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests — honor Retry-After
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds to wait }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Timeout:
      description: Upstream LLM timeout (safe to retry once)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Source:
      type: string
      enum: [mobile]
      default: mobile
      description: Always "mobile" for the mobile SDK.
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    ChatMessageRequest:
      type: object
      required: [projectId, message, visitorId]
      properties:
        projectId: { type: string, format: uuid }
        message: { type: string, maxLength: 2000 }
        visitorId: { type: string, maxLength: 100 }
        source: { $ref: "#/components/schemas/Source" }
        sessionId:
          type: string
          format: uuid
          nullable: true
          description: Omit/null for the first message; send to continue a thread.
        conversationHistory:
          type: array
          description: Recent turns for context (last ~10).
          items:
            type: object
            properties:
              role: { type: string, enum: [user, assistant] }
              content: { type: string }
        context:
          type: object
          description: Optional analytics metadata.
          properties:
            device: { type: string }
            os: { type: string }
            osVersion: { type: string }
            timezone: { type: string }
            language: { type: string }
            appVersion: { type: string }
          additionalProperties: true
    ChatMessageResponse:
      type: object
      properties:
        response: { type: string, description: Assistant reply; may be empty during handoff. }
        sessionId: { type: string, format: uuid, description: Persist this. }
        sources:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              relevance: { type: number }
        toolCalls:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              success: { type: boolean }
              duration: { type: number }
        processingTime: { type: number }
        requestId: { type: string }
        handoff:
          type: object
          description: Present when the AI escalated or the thread is agent/queue-owned.
          properties:
            triggered: { type: boolean }
            reason:
              type: string
              description: e.g. in_queue, agent_handling, customer_request
            queuePosition: { type: integer }
            estimatedWait: { type: string }
            conversationId: { type: string, format: uuid }
...